requirements.txt
PySide6>=6.5.0
requests>=2.31.0
pandas>=2.0.0
mplfinance>=0.12.9
matplotlib>=3.7.0
python-dotenv>=1.0.0
qt-material>=2.14
pydantic>=2.0.0